[+] Add soems lentgh di

+++ my_printf/formats/a_format.c
@@ -19,15 +19,15 @@ static int call_di_format(char **buffer, ...)
     return 0;
 }
 
-static int formating_buffer(char **buffer, int nbr, my_flags_t *flgs)
+static int formating_buffer(char **buffer, double nbr, my_flags_t *flgs)
 {
     int len;
 
     my_revstr(*buffer);
     add_buffer(buffer, "x0", 2);
-    (nbr < 0 || nbr == -0.) && add_buffer(buffer, "-", 1);
-    nbr >= 0 && flgs->has_plus && add_buffer(buffer, "+", 1);
-    nbr >= 0 && !flgs->has_plus && flgs->has_space &&
+    (nbr < 0. || nbr == -0.) && add_buffer(buffer, "-", 1);
+    nbr >= 0. && flgs->has_plus && add_buffer(buffer, "+", 1);
+    (nbr >= 0.) && !(flgs->has_plus) && flgs->has_space &&
         add_buffer(buffer, " ", 1);
     my_revstr(*buffer);
     len = my_strlen(*buffer);
